warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@11928 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/VKontakte/src/vk_files.cpp
@@ -230,7 +230,7 @@ void CVkProto::OnReciveUploadServer(NETLIBHTTPREQUEST *reply, AsyncHttpRequest *
 	pUploadReq->pData = pData;
 
 	pData += DataBegin.GetLength();
-	long lBytes = fread(pData, 1, iFileLen, pFile);
+	long lBytes = (long)fread(pData, 1, iFileLen, pFile);
 	fclose(pFile);
 
 	if (lBytes != iFileLen) {
